backend/celebrity/requirements.txt
python-dotenv==0.13.0
dnspython==2.0.0
kafka-python==2.0.1
redis==3.5.3
SQLAlchemy==1.3.19
psycopg2==2.8.6
Flask-SQLAlchemy==2.4.4
requests==2.18.4
PyJWT==1.7.1